dataanalysis_exercise.
created dataanalysis-exercise folder within portfolio. so far this folder contains a data folder and a code folder, and a dataanalysis_excercise.qmd file that stores the first part of the assignment. In this first part, you cleaned a dataset from the CDC and your group member should be able to follow using the .qmd file. The code file contains a processing code file that contains the processing code. Note, to pasted all of the processing code into the dataanalysis_excercise.qmd file.