Data Analysis Exercise
added the data analysis exercise to the website. Key notes, when adding a new menu you must update the quarto.yml file. The text section is the title of the new page. the href is the path extension to the website.  When editing the exercise document the title must match the text section in the quarto.yml file. OH, and the href section in the quarto.yml file must match the name of the qmd file that you  are editing that contains the exercise.